+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.3.3",
+    "canvasjs": "^1.8.3",
     "chart.js": "^4.4.3",
     "chartjs-adapter-moment": "^1.0.1",
     "express": "^4.19.2",
@@ -20,6 +21,7 @@
     "moment": "^2.30.1",
     "mui": "^0.0.1",
     "react": "^18.3.1",
+    "react-bootstrap": "^2.10.4",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.3.1",
     "react-gauge-component": "^1.2.21",
