+++ src/App.js
@@ -4,26 +4,45 @@
 //Ali Elgazzar comme
 //import React, { useState, useEffect } from 'react';
 //import io from 'socket.io-client';
-import VibPage from './VibPage';
-import Home from './Home';
+import VibPage from './Pages/VibPage';
+import Home from './Pages/Home';
 import Login from './Login/Login'
-import EnergyMonitoring from './EnergyMonitoring';
-import MPMonitoring from './MPMonitoring';
-import EnergyAlarm from './EnergyAlarm'
+import EnergyMonitoring from './Pages/EnergyMonitoring';
+import MPMonitoring from './Pages/MPMonitoring';
+import EnergyAlarm from './Pages/EnergyAlarm'
 import SignUp from './Login/SignUp';
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
-import PredictiveAlarm from './PredictiveAlarm';
+import PredictiveAlarm from './Pages/PredictiveAlarm';
 
-// class Start extends Comment {
-//   constructor(props) {
-//     super(props);
-//   }
-//   render() {
-//     return (
-      
-//     );
-//   }
-// }
+
+document.addEventListener("contextmenu", function (event) {
+  alert("Inspect Elements Not Allowed");
+  event.preventDefault();
+})
+document.addEventListener("keydown", function (event) {
+  if (event.key === "F12" ||
+      (event.ctrlKey && event.shiftKey && event.key === "I") ||
+      (event.ctrlKey && event.shiftKey && event.key === "J") ||
+      (event.ctrlKey && event.key === "U")) {
+      alert("Inspect Elements Not Allowed");
+      event.preventDefault();
+  }
+})
+
+var devtoolsOpen = false;
+var threshold = 160;
+setInterval(function () {
+  if ((window.outerWidth - window.innerWidth) > threshold ||
+      (window.outerHeight - window.innerHeight) > threshold) {
+      if (!devtoolsOpen) {
+          devtoolsOpen = true;
+          alert("Developer Tools Detected. Please close them.");
+      }
+  } else {
+      devtoolsOpen = false;
+  }
+}, 1000)
+
 
 const App = () => {
